Move: porduct class to itertoolsext
